Added new methods and refactored Command classes
In this commit, several changes were made to improve the structure and functionality of the Command classes.

In `CommandDefinition.cs`, the `QuiCLI.Common` namespace was added and three new methods were introduced: `FromMethod`, `GetArguments`, and `ConvertCamelCaseToParameterName`. These methods respectively create a `CommandDefinition` from a `MethodInfo` object, generate a list of `ArgumentDefinition` objects, and convert a camel case string to a parameter name.

`CommandGroup.cs` was refactored to use the new `FromMethod` method from `CommandDefinition.cs`, and the `GetArguments` and `ConvertCamelCaseToParameterName` methods were removed as they are now part of the `CommandDefinition` class.